feat: package refactor
