+++ formatter/src/main/kotlin/edu/austral/ingsis/gradle/formatter/RuleParser.kt
@@ -1,4 +1,4 @@
-package edu.austral.ingsis.gradle.formatter.util
+package edu.austral.ingsis.gradle.formatter
 
 import kotlinx.serialization.json.Json
 import java.io.File
